:zap: buildステージ修正

+++ function/create-redirect-config.js
@@ -11,8 +11,8 @@ const pipeline = util.promisify(stream.pipeline);
 
 const client = new S3Client({});
 const bucketName = process.env.REDIRECT_CONFIG_BUCKET;
-let singleRedirectRules = [];
-let regexRedirectRules = [];
+let singleRedirectRules = {};
+let regexRedirectRules = {};
 
 // S3から各WPのリダイレクト設定を取得してまとめる
 async function createRedirectConfig() {
@@ -28,37 +28,40 @@ async function createRedirectConfig() {
 
         const listCommand = new ListObjectsV2Command(listParams);
         const listResponse = await client.send(listCommand);
-        console.log(listResponse);
-        for (const content of listResponse.Contents) {
-            if (content.Key.endsWith('regex-redirect-config.json')) {
-                isRegex = true;
-            } else if (content.Key.endsWith('redirect-config.json')) {
-                isRegex = true;
-            } else {
-                continue;
+
+        if ('Contents' in listResponse) {
+            for (const content of listResponse.Contents) {
+                if (content.Key.endsWith('regex-redirect-config.json')) {
+                    isRegex = true;
+                } else if (content.Key.endsWith('redirect-config.json')) {
+                    isRegex = true;
+                } else {
+                    continue;
+                }
+
+                const getObjectParams = {
+                    Bucket: bucketName,
+                    Key: content.Key,
+                }
+                const getObjectCommand = new GetObjectCommand(getObjectParams);
+                const { Body } = await client.send(getObjectCommand);
+
+                // ストリームを使用してデータを読み込み、一時ファイルを作成する
+                const tempFilePath = `temp-${Date.now()}.json`;
+                await pipeline(Body, fs.createWriteStream(tempFilePath));
+
+                const fileData = JSON.parse(fs.readFileSync(tempFilePath));
+
+                if (isRegex) {
+                    regexRedirectRules = { ...regexRedirectRules, ...fileData};
+                } else {
+                    singleRedirectRules = { ...singleRedirectRules, ...fileData};
+                }
+
+                fs.unlinkSync(tempFilePath);
             }
-
-            const getObjectParams = {
-                Bucket: bucketName,
-                Key: content.Key,
-            }
-            const getObjectCommand = new GetObjectCommand(getObjectParams);
-            const { Body } = await client.send(getObjectCommand);
-
-            // ストリームを使用してデータを読み込み、一時ファイルを作成する
-            const tempFilePath = `temp-${Date.now()}.json`;
-            await pipeline(Body, fs.createWriteStream(tempFilePath));
-
-            const fileData = JSON.parse(fs.readFileSync(tempFilePath));
-
-            if (isRegex) {
-                regexRedirectRules.concat(fileData);
-            } else {
-                singleRedirectRules.concat(fileData);
-            }
-
-            fs.unlinkSync(tempFilePath);
         }
+
 
         continuationToken = listResponse.NextContinuationToken;
         if (!continuationToken) {
@@ -72,4 +75,5 @@ async function createRedirectConfig() {
 
 createRedirectConfig().catch((error) => {
     console.error(error);
+    process.exit(1);
 });
